test: share dictionary temp path fixtures
