+++ test/dictionaries/wiktionary/test_wiktionary_dictionary_service.py
@@ -5,35 +5,12 @@
 from __future__ import annotations
 
 import json
-from collections.abc import Generator
 from pathlib import Path
 
 import requests
-from pytest import MonkeyPatch, fixture, raises
+from pytest import MonkeyPatch, raises
 
-from scinoephile.common.file import get_temp_directory_path, get_temp_file_path
 from scinoephile.dictionaries.wiktionary import WiktionaryDictionaryService
-
-
-@fixture
-def local_data_dir_path() -> Generator[Path]:
-    """Provide a temporary canonical local data directory."""
-    with get_temp_directory_path() as dir_path:
-        yield dir_path
-
-
-@fixture
-def runtime_data_dir_path() -> Generator[Path]:
-    """Provide a temporary runtime canonical data directory."""
-    with get_temp_directory_path() as dir_path:
-        yield dir_path
-
-
-@fixture
-def database_path() -> Generator[Path]:
-    """Provide a temporary SQLite database path."""
-    with get_temp_file_path(".db") as temp_path:
-        yield temp_path
 
 
 def _write_fixture_jsonl(jsonl_path: Path):
